Use fake $prefix to avoid symlink farm indirection for rootdir
Since everything is built as a Meson subproject now, $prefix no longer
needs to be a real path that packages read from (e.g. for pkg-config
include and library paths), but we still need to make sure there's nothing
in it (so dependencies always use subproject overrides instead of binaries
from previous builds).  Convert $prefix to a fake path and use --destdir
to install into the right place, avoiding the need for a symlink farm.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>