workflows: periodically check for updates
Automatically check for updates once a week, whenever wraps are updated,
or on demand.  Have openslide-bot maintain a PR with Meson wrap updates,
and an issue tracking any remaining updates reported by
`./build.sh updates`.

We need to do this in the context of openslide-bot, rather than the
GITHUB_TOKEN, because PRs created by the GITHUB_TOKEN a) can't use a head
branch in a fork, and b) don't automatically run CI.  However, we want to
label affected issues and PRs, which would require giving openslide-bot
permissions.  Instead, we have a second, privileged workflow to detect
and label these issues/PRs.

This workflow maintains the PR branch in openslide-bot's fork of
openslide-winbuild.  Since the fork may have a stale main branch with
different Actions workflows, the openslide-bot token must have the
`workflow` scope to allow pushing the current workflows to the branch.
Workflows have been disabled in that repo to minimize the effect.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>