Postprocess binaries in Meson
Create split debuginfo and check exported symbols in Meson rather than in
the build.sh wrapper.  Support all three platforms.  On Linux and macOS,
also update the slidetool RPATH.

To do this, the Meson custom_target needs to be able to find the slidetool
executable in the build directory.  Backport an OpenSlide commit to store
the exe object in a Meson variable.

Also add a Meson rule to copy openslide.jar to the artifacts directory,
so build.sh only has to look for output artifacts in one place.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>