Generate source archive with `meson dist`
Reuse existing Meson functionality rather than hardcoding a list of files
to include in the archive.  As a consequence, rather than shipping a
nested compressed tarball for each dependency, ship the actual unpacked
source trees.

Meson supports tar.xz, tar.gz, and ZIP.  The sdist archive no longer needs
to be natively supported by Windows, since Windows builds now always use
a Linux container.  ZIP is less efficient because each file is compressed
individually, so avoid it.  tar.xz would be ideal, but PyPI requires
tar.gz, and it would be nice to reuse the same sdist archive for eventual
Python wheel support.  The source archive isn't expected to be widely
used, so we avoid generating both tar.gz and tar.xz archives, and accept
the lower compression ratio.

Pin the currently configured version when generating a source archive.
When building from a source archive, allow overriding the version in the
usual way, but if not overridden, append "-local" to make it clear that
the build isn't an official image.

Normally, each platform build only enables the dependencies that are
needed on that platform.  The Windows builds currently enable all
dependencies, but that may not always be the case.  Add a Meson option,
for use when generating sdist archives, that enables all dependencies so
they all end up in the tarball.  In addition, add a fully-populated
VERSIONS.md to the sdist archive so there's a single record of all
dependency versions.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>